Move debug window specific code to a partial class
